BoxingForJustice-main/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
openai==1.68.2
python-dotenv==1.0.1
pydantic==2.10.6
numpy==2.2.4
